requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
absl-py=2.2.0=pypi_0
anyio=4.9.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.5=pypi_0
attrs=25.3.0=pypi_0
babel=2.17.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
blas=1.0=mkl
bleach=6.2.0=pypi_0
brotli-python=1.0.9=py39h5da7b33_9
bzip2=1.0.8=h8ffe710_4
ca-certificates=2025.1.31=h56e8100_0
cairo=1.16.0=he04af86_2
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py39ha55e580_0
chardet=3.0.4=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.1.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cpuonly=2.0=0
cycler=0.12.1=pypi_0
debugpy=1.8.13=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
eigen=3.4.0=h2d74725_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.1=pypi_0
ffmpeg=4.3.1=ha925a31_0
filelock=3.13.1=py39haa95532_0
flatbuffers=25.2.10=pypi_0
fonttools=4.51.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py39h827c3e9_0
graphite2=1.3.14=hd77b12b_1
graphviz=0.8.4=pypi_0
gst-plugins-base=1.22.3=hfc3ed34_1
gstreamer=1.22.3=h1779075_1
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h11=0.14.0=pypi_0
h2=4.2.0=pyhd8ed1ab_0
h5py=3.13.0=pypi_0
harfbuzz=10.2.0=hf458a27_0
hdf5=1.14.5=ha36df97_2
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.10=pyhd8ed1ab_1
imbalanced-learn=0.12.4=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.5=pypi_0
ipython=8.18.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=py39haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
json5=0.12.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.6=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
keras=3.9.0=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h5b6d351_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libabseil=20240116.2=cxx17_h5da7b33_0
libblas=3.9.0=26_win64_mkl
libcblas=3.9.0=26_win64_mkl
libclang=18.1.1=pypi_0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libglib=2.78.4=ha17d25a_0
libiconv=1.16=h2bbff1b_3
libjpeg-turbo=2.0.0=h196d8e1_0
liblapack=3.9.0=26_win64_mkl
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.20=h70ee33d_0
libprotobuf=4.25.3=hf2fb9eb_0
libtiff=4.5.1=h44ae7cf_1
libuv=1.48.0=h827c3e9_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py39h827c3e9_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.1.3=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h827c3e9_2
mkl_fft=1.3.11=py39h827c3e9_0
mkl_random=1.2.8=py39hc64d2fc_0
ml-dtypes=0.5.1=pypi_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py39haa95532_0
mxnet=1.7.0.post2=pypi_0
namex=0.0.8=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=py39haa95532_0
notebook=7.3.3=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.21.6=pypi_0
opencv=4.10.0=py39h545ae1a_3
openjpeg=2.5.2=hae555c5_0
openpyxl=3.1.2=pypi_0
openssl=3.4.1=ha4e3fda_0
opt-einsum=3.4.0=pypi_0
optree=0.14.1=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
patsy=0.5.6=pypi_0
pcre2=10.42=h0ff8eda_1
pillow=11.1.0=py39h096bfcc_0
pip=25.0=py39haa95532_0
pixman=0.40.0=h2bbff1b_1
platformdirs=4.3.7=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pypi_0
protobuf=5.29.4=pypi_0
psutil=7.0.0=pypi_0
psycopg2=2.9.9=py39h2bbff1b_0
pure-eval=0.2.3=pypi_0
pyaml=25.7.0=pypi_0
pycparser=2.22=pyh29332c3_1
pygments=2.19.1=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.18=h1aa4202_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=3.3.0=pypi_0
python_abi=3.9=2_cp39
pytorch=2.2.0=py3.9_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=pypi_0
pywin32=310=pypi_0
pywinpty=2.0.15=pypi_0
pyyaml=6.0.2=py39h827c3e9_0
pyzmq=26.4.0=pypi_0
qtbase=6.7.2=h0804d20_0
referencing=0.36.2=pypi_0
requests=2.32.3=pyhd8ed1ab_1
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.24.0=pypi_0
scikit-learn=1.4.2=pypi_0
scikit-optimize=0.10.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=72.1.0=py39haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.2=pypi_0
sympy=1.13.3=py39haa95532_1
tbb=2021.8.0=h59b6b97_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.5.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.4.0=pypi_0
tomli=2.2.1=pypi_0
torchaudio=2.2.0=py39_cpu
torchvision=0.17.0=py39_cpu
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing_extensions=4.12.2=py39haa95532_0
tzdata=2024.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
uri-template=1.3.0=pypi_0
urllib3=1.22=pypi_0
vc=14.42=haa95532_4
vc14_runtime=14.42.34438=hfd919c2_24
vs2015_runtime=14.42.34438=h7142326_24
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.17.2=pypi_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h8cc25b3_1
zstandard=0.23.0=py39ha55e580_1
zstd=1.5.6=h8880b57_0
